Add Vegas example for all three tasks
